A classical approach to solving the cook membrane problem.
In this work we take a classical approach to solving the equations
governing quasi-static finite-strain compressible elasticity, with code
based on `step-44`. The formulation adopted here is that seen in many
texts on solid mechanics and can be used as the starting point for
extension into many topics such as material anisotropy, rate dependence
or plasticity, or even as a component of multi-physics problems.