Update DataOut filter interface usage.
As far as I can see, all these filters in this code gallery program do
is select the locally owned cells. But this is what DataOut has been
doing automatically for a long time already, so the filters are no longer
necessary.

The face filter class was not even used :-)